[#4] Rename packages
