+++ containers-contrib/containers-contrib.cabal
@@ -1,7 +1,7 @@
-name:                map-contrib
+name:                containers-contrib
 version:             0.0.0
-synopsis:            Different functions implemented via general @map-sig@ interface
-description:         Different functions implemented via general @map-sig@ interface
+synopsis:            Different functions implemented via general containers-sig interface
+description:         Different functions implemented via general containers-sig interface
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -23,6 +23,9 @@ library
   hs-source-dirs:      src
   exposed-modules:     Map.Contrib.Dummy
                        Map.Contrib.Group
-  build-depends:       base, map-sig, map-sig-extended
+
+  build-depends:       base
+                     , containers-sig
+                     , containers-sig-readonly
   ghc-options:         -Wall -fhide-source-paths
   default-language:    Haskell2010
