+++ containers-contrib/src/Map/Contrib/Dummy.hs
