+++ containers-example/containers-example.cabal
@@ -1,7 +1,7 @@
-name:                map-example
+name:                containers-example
 version:             0.0.0
-synopsis:            Backpack implementation of @map-sig@
-description:         Backpack implementation of @map-sig@
+synopsis:            Usage example of containers-sig
+description:         Usage example of containers-sig
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -23,16 +23,17 @@ executable map-exe
   hs-source-dirs:      app
   main-is:             Main.hs
   build-depends:       base
-                     , map-containers-intmap
-                     , map-containers-map
-                     , map-unordered
-                     , map-contrib
-  mixins:              map-contrib (Map.Contrib.Group as Map.Contrib.Group.IntMap)
-                          requires (Map as Map.IntMap)
-                     , map-contrib (Map.Contrib.Group as Map.Contrib.Group.Map)
-                          requires (Map as Map.Map)
-                     , map-contrib (Map.Contrib.Group as Map.Contrib.Group.HashMap)
-                          requires (Map as Map.HashMap)
+                     , containers-int-strict
+                     , containers-ordered-strict
+                     , containers-unordered-strict
+                     , containers-contrib
+
+  mixins:              containers-contrib (Map.Contrib.Group as Map.Contrib.Group.Int)
+                                 requires (Map as Map.Int)
+                     , containers-contrib (Map.Contrib.Group as Map.Contrib.Group.Ord)
+                                 requires (Map as Map.Ord)
+                     , containers-contrib (Map.Contrib.Group as Map.Contrib.Group.Hash)
+                                 requires (Map as Map.Hash)
 
   ghc-options:         -Wall -threaded -rtsopts -with-rtsopts=-N -fhide-source-paths
   default-language:    Haskell2010
