+++ containers-int-strict/LICENSE
