+++ containers-int-strict/containers-int-strict.cabal
@@ -1,7 +1,7 @@
-name:                map-containers-map
+name:                containers-int-strict
 version:             0.0.0
-synopsis:            Backpack implementation of @map-sig@
-description:         Backpack implementation of @map-sig@
+synopsis:            Backpack implementation of containers-sig for Data.IntMap.Strict
+description:         Backpack implementation of containers-sig for Data.IntMap.Strict
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -22,9 +22,9 @@ source-repository head
 library
   hs-source-dirs:      src
 
-  exposed-modules:     Map.Map
-  reexported-modules:  Map.Map as Map
+  exposed-modules:     Map.Int
+  reexported-modules:  Map.Int as Map
 
   build-depends:       base, containers
   ghc-options:         -Wall -fhide-source-paths
-  default-language:   Haskell2010
+  default-language:    Haskell2010
