+++ containers-ordered-strict/containers-ordered-strict.cabal
@@ -1,7 +1,7 @@
-name:                map-sig
+name:                containers-ordered-strict
 version:             0.0.0
-synopsis:            Backpack signatures for Map-like interfaces
-description:         Backpack signatures for Map-like interfaces
+synopsis:            Backpack implementation of containers-sig for Data.Map.Strict
+description:         Backpack implementation of containers-sig for Data.Map.Strict
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -21,7 +21,10 @@ source-repository head
 
 library
   hs-source-dirs:      src
-  signatures:          Map
-  build-depends:       base
+
+  exposed-modules:     Map.Ord
+  reexported-modules:  Map.Ord as Map
+
+  build-depends:       base, containers
   ghc-options:         -Wall -fhide-source-paths
-  default-language:   Haskell2010
+  default-language:    Haskell2010
