+++ containers-sig-readonly/containers-sig-readonly.cabal
@@ -1,7 +1,7 @@
-name:                map-sig-extended
+name:                containers-sig-readonly
 version:             0.0.0
-synopsis:            Backpack signatures for Map-like interfaces
-description:         Backpack signatures for Map-like interfaces
+synopsis:            Backpack signatures for readonly Map-like interfaces
+description:         Backpack signatures for readonly Map-like interfaces
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -22,6 +22,6 @@ source-repository head
 library
   hs-source-dirs:      src
   signatures:          Map
-  build-depends:       base, map-sig
+  build-depends:       base
   ghc-options:         -Wall -fhide-source-paths
-  default-language:   Haskell2010
+  default-language:    Haskell2010
