+++ containers-sig/CHANGELOG.md
@@ -1,7 +1,7 @@
 Change log
 ==========
 
-`map-containers` uses [PVP Versioning][1].
+`containers-sig` uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
 0.0.0
