+++ containers-sig/containers-sig.cabal
@@ -1,7 +1,7 @@
-name:                map-unordered
+name:                containers-sig
 version:             0.0.0
-synopsis:            Backpack implementation of @map-sig@
-description:         Backpack implementation of @map-sig@
+synopsis:            Backpack signatures for Map-like interfaces
+description:         Backpack signatures for Map-like interfaces
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -21,13 +21,7 @@ source-repository head
 
 library
   hs-source-dirs:      src
-
-  exposed-modules:     Map.HashMap
-  reexported-modules:  Map.HashMap as Map
-
+  signatures:          Map
   build-depends:       base
-                     , hashable
-                     , unordered-containers
-
   ghc-options:         -Wall -fhide-source-paths
-  default-language:   Haskell2010
+  default-language:    Haskell2010
