+++ containers-unordered-strict/containers-unordered-strict.cabal
@@ -1,7 +1,7 @@
-name:                map-containers-intmap
+name:                containers-unordered-strict
 version:             0.0.0
-synopsis:            Backpack implementation of @map-sig@
-description:         Backpack implementation of @map-sig@
+synopsis:            Backpack implementation of containers-backpack for Data.HashMap.Strict
+description:         Backpack implementation of containers-backpack for Data.HashMap.Strict
 homepage:            https://github.com/kowainik/containers-backpack
 bug-reports:         https://github.com/kowainik/containers-backpack/issues
 license:             MPL-2.0
@@ -22,9 +22,12 @@ source-repository head
 library
   hs-source-dirs:      src
 
-  exposed-modules:     Map.IntMap
-  reexported-modules:  Map.IntMap as Map
+  exposed-modules:     Map.Hash
+  reexported-modules:  Map.Hash as Map
 
-  build-depends:       base, containers
+  build-depends:       base
+                     , hashable
+                     , unordered-containers
+
   ghc-options:         -Wall -fhide-source-paths
-  default-language:   Haskell2010
+  default-language:    Haskell2010
